+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 1.5.5
+
 # 1.5.4
 - Increase Oddion Hitbox Size
 - Decrease Pumpkin Warden Hitbox Size